requirements.txt
# FastAPI and Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
supabase>=2.10.0

# AI and ML Services
google-cloud-vision>=3.8.0
google-generativeai>=0.8.0
groq>=0.11.0

# Environment and Configuration
python-dotenv>=1.0.1

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Image Processing
Pillow>=11.0.0

# HTTP Requests
requests>=2.32.0
httpx>=0.27.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.9.0

# Logging and Monitoring
loguru>=0.7.2

# Type Checking
typing-extensions>=4.12.0